requirements.txt
Django>=3.1.5
PyYAML>=5.4.1
djangorestframework>=3.12.2
psycopg2>=2.8.6
uritemplate>=3.0.1
